Moving TruncateMaxLines to main components folder
