+++ src/components/truncate-max-lines/types.ts
